requirements.txt
Flask~=3.1.0
Pillow~=11.1.0
pytesseract~=0.3.13
Werkzeug~=3.1.3
PyMonad~=2.4.0